+++ src/plugins/healthcheck/public/components/services/health.ts
@@ -5,13 +5,13 @@
 
 import { TaskInfo } from 'src/core/common/healthcheck';
 
-const mapStatusHealth = {
+const taskStatusColorMapping = {
   green: 'success',
   yellow: 'warning',
   red: 'danger',
   gray: 'subdued',
 };
 
-export function getHealthFromStatus(status: TaskInfo['result']) {
-  return (status && mapStatusHealth[status]) || mapStatusHealth.gray;
+export function mapTaskStatusToHealthColor(status: TaskInfo['result']) {
+  return (status && taskStatusColorMapping[status]) || taskStatusColorMapping.gray;
 }