+++ CHANGELOG.md
@@ -7,8 +7,10 @@ All notable changes to the Wazuh app project will be documented in this file.
 ### Added
 
 - Support for Wazuh 4.14.0
+- Added Observability and Anomaly Detection plugins to default installation [#843](https://github.com/wazuh/wazuh-dashboard/pull/843)
+- Added ML Commons plugin to default installation [#875](https://github.com/wazuh/wazuh-dashboard/pull/875)
 
-## Wazuh dashboard v4.13.0 - OpenSearch Dashboards 2.19.2 - Revision 05
+## Wazuh dashboard v4.13.0 - OpenSearch Dashboards 2.19.2 - Revision 07
 
 ### Added