+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to the Wazuh app project will be documented in this file.
 ### Added
 
 - Support for Wazuh 5.0.0
+- Health check service [#811](https://github.com/wazuh/wazuh-dashboard/pull/811) [#866](https://github.com/wazuh/wazuh-dashboard/pull/866)
+- Added Health Check plugin [#870](https://github.com/wazuh/wazuh-dashboard/pull/870)
 
 ### Removed