+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Added ML Commons plugin to default installation [#875](https://github.com/wazuh/wazuh-dashboard/pull/875)
 - Added Assistant Dashboards plugin to default installation [#821](https://github.com/wazuh/wazuh-dashboard/pull/821)
 
-## Wazuh dashboard v4.13.0 - OpenSearch Dashboards 2.19.2 - Revision 07
+## Wazuh dashboard v4.13.0 - OpenSearch Dashboards 2.19.2 - Revision 08
 
 ### Added