+++ CHANGELOG.md
@@ -18,6 +18,12 @@ All notable changes to the Wazuh app project will be documented in this file.
 - Added Assistant Dashboards plugin to default installation [#821](https://github.com/wazuh/wazuh-dashboard/pull/821)
 - Added directive to the dashboard service to automatically restart it on crash [#919](https://github.com/wazuh/wazuh-dashboard/pull/919)
 
+## Wazuh dashboard v4.13.1 - OpenSearch Dashboards 2.19.2 - Revision 01
+
+### Added
+
+- Support for Wazuh 4.13.1
+
 ## Wazuh dashboard v4.13.0 - OpenSearch Dashboards 2.19.2 - Revision 08
 
 ### Added